Feat: Post 삭제 API 수정

+++ server/api/post.js
@@ -8,7 +8,7 @@ router.get("/", (req, res) => {
   const idx = req.query["post_id"];
   pool.getConnection((err, connection) => {
     const sql = `SELECT id, title, writer, reg_date, comment_cnt, view_cnt, like_cnt, content  from POST where (id=${idx})`;
-    
+
     // + comment 관련 데이터랑 해시태그 관련 데이터도 추가해야 함
     connection.query(sql, (err, rows) => {
       res.json(rows[0]);
@@ -46,11 +46,23 @@ router.put("/", (req, res) => {
 });
 
 router.delete("/", (req, res) => {
+  const password = crypto
+    .createHash("sha256")
+    .update(req.body.password)
+    .digest("base64");
+
   const con = pool.getConnection((err, connection) => {
-    const sql = `delete from post where (id=${req.body.id})`;
+    const sql = `select password from post where (id=${req.body.id})`
     connection.query(sql, (err, rows) => {
-      if (err) res.send(err);
-      else res.status(200).send("Post has deleted");
+      if (rows[0].password == password) {
+        const sql2 = `delete from post where (id=${req.body.id})`;
+        connection.query(sql2, (err, rows) => {
+          if (err) res.send(err);
+          else res.status(200).send("Post has deleted");
+        })
+      } else {
+        res.status(202).send("wrong password")
+      }
     });
     connection.release();
   });
